# Introduce a strategy that improves pack(LHS) using transform dialect.