--- a/compiler/src/iree/compiler/Codegen/TransformDialectStrategies/CPU/PackStrategy.cpp
+++ b/compiler/src/iree/compiler/Codegen/TransformDialectStrategies/CPU/PackStrategy.cpp
@@ -10,10 +10,10 @@
 #include "iree/compiler/Codegen/LLVMCPU/TransformExtensions/LLVMCPUExtensions.h"
 #include "iree/compiler/Codegen/Passes.h"
 #include "iree/compiler/Codegen/TransformDialectStrategies/CPU/Common.h"
-#include "iree/compiler/Codegen/TransformDialectStrategies/Common/AbstractReductionStrategy.h"
 #include "iree/compiler/Codegen/TransformDialectStrategies/Common/Common.h"
 #include "llvm/Support/Debug.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/GPU/IR/GPUDialect.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
 #include "mlir/Dialect/Transform/IR/TransformOps.h"
 #include "mlir/Dialect/Vector/Transforms/VectorRewritePatterns.h"
@@ -28,57 +28,71 @@ namespace iree_compiler {
 namespace cpu {
 
 using IREE::transform_dialect::ApplyPatternsOpPatterns;
-using IREE::transform_dialect::TileToForallAndWorkgroupCountRegionOp;
-using transform::LowerPackOp;
 using transform::MatchOp;
 
+PackStrategy PackStrategy::create(MLIRContext *context, tensor::PackOp packOp,
+                                  bool lowerToAVX2) {
+  LLVM_DEBUG(DBGS() << "use CPU pack strategy\n");
+  PackStrategy strategy;
+  strategy.numThreads = 1;
+  strategy.lowerToAVX2 = lowerToAVX2;
+
+  auto destType = packOp.getDestType();
+  int rank = destType.getRank();
+
+  strategy.tileSizes.append(rank, 1);
+  auto getMaxDivisor = [](int64_t val, int64_t tileSize) -> int64_t {
+    if (ShapedType::isDynamic(val)) return tileSize;
+    FailureOr<int64_t> res = maxDivisorOfValueBelowLimit(val, tileSize);
+    if (failed(res)) return 1;
+    return res.value();
+  };
+
+  for (int i = rank - 1, cnt = 0; i >= 0; --i) {
+    int64_t size = destType.getDimSize(i);
+    if (size == 1) continue;
+    strategy.tileSizes[i] = getMaxDivisor(size, 8);
+    cnt++;
+    if (cnt >= 2) break;
+  }
+
+  return strategy;
+}
+
 void buildPackStrategy(ImplicitLocOpBuilder &b, Value variantH,
-                       const PackConfig &config) {
+                       const PackStrategy &strategy) {
   MLIRContext *ctx = b.getContext();
   Value packOp = b.create<transform::MatchOp>(
       variantH, tensor::PackOp::getOperationName());
 
-  // TODO(hanchung): Support multi-threaded.
-  SmallVector<int64_t> numThreads = {1};
-  auto tileToForAll = b.create<TileToForallAndWorkgroupCountRegionOp>(
-      packOp, numThreads, transform::NumThreadsSpec(),
-      b.getArrayAttr(
-          mlir::gpu::GPUBlockMappingAttr::get(ctx, mlir::gpu::Blocks::DimX)));
+  TileToForallAndFuseAndDistributeResult distResult =
+      buildTileFuseDistToForallAndWorgroupCountWithNumThreads(
+          b, variantH, packOp, /*opsHToFuse=*/ValueRange(),
+          /*numThreads=*/
+          getAsOpFoldResult(b.getIndexArrayAttr({strategy.getNumThreads()})),
+          b.getArrayAttr(mlir::gpu::GPUBlockMappingAttr::get(
+              ctx, mlir::gpu::Blocks::DimX)));
 
-  // Lower to pad + expand_reshape + transpose
-  packOp = tileToForAll.getTiledOp();
-  auto lowerPackOp = b.create<LowerPackOp>(
-      transform::OperationType::get(ctx, tensor::PadOp::getOperationName()),
-      transform::OperationType::get(ctx,
-                                    tensor::ExpandShapeOp::getOperationName()),
-      transform::OperationType::get(ctx,
-                                    linalg::TransposeOp::getOperationName()),
-      b.create<transform::CastOp>(transform::OperationType::get(
-                                      ctx, tensor::PackOp::getOperationName()),
-                                  packOp));
-  Value transposeOp = lowerPackOp.getTransposeOp();
+  // Lower pack op, generalize the transpose op and fold unit dims away.
+  Value transposeOp = std::get<2>(buildLowerPack(b, distResult.tiledOpH));
   SmallVector<OpFoldResult> tileSizes =
-      getAsOpFoldResult(b.getIndexArrayAttr({1, 8, 8}));
-  auto tiletoScfForOp = b.create<transform::TileOp>(transposeOp, tileSizes);
-
-  // Generalize the transpose op and fold unit dims away.
-  transposeOp = tiletoScfForOp.getTiledLinalgOp();
+      getAsOpFoldResult(b.getIndexArrayAttr(strategy.getTransposeTileSizes()));
+  TileToScfForAndFuseResult tileResult =
+      buildTileFuseToScfFor(b, variantH, transposeOp, ValueRange(), tileSizes);
+  transposeOp = tileResult.tiledOpH;
   auto pdlOpType = pdl::OperationType::get(ctx);
   b.create<transform::GeneralizeOp>(
       pdlOpType, b.create<transform::CastOp>(pdlOpType, transposeOp));
 
   ApplyPatternsOpPatterns configuration;
-  configuration.rankReducingLinalgViaReshapes = true;
-  Value funcH =
-      b.create<transform::MatchOp>(variantH, func::FuncOp::getOperationName());
-  funcH = mlir::iree_compiler::buildCanonicalizationAndEnablingTransforms(
-      b, configuration, funcH);
+  configuration.rankReducingLinalg = true;
+  buildCanonicalizationAndEnablingTransforms(b, configuration, variantH);
 
   vector::LowerVectorsOptions lowerVectorsOptions;
   lowerVectorsOptions
       .setVectorTransferSplit(vector::VectorTransferSplit::LinalgCopy)
       .setVectorTransposeLowering(vector::VectorTransposeLowering::Shuffle)
-      .setTransposeAVX2Lowering(config.lowerToAVX2)
+      .setTransposeAVX2Lowering(strategy.getLowerToAVX2())
       .setUnrollVectorTransfers(true);
   buildCommonTrailingStrategy(b, variantH, lowerVectorsOptions);
 }