It starts with single-threaded approach on pack op w/o padding semantics. The flow of the strategy is

1. Lower the pack op into pad + reshape + transpose
2. Tile transpose op and fold unit dims to generate a tiled transpose op with tensor<8x8xf32> shape
3. Vectorize all the operations.
4. Bufferization
5. Kick in vector lowering and enable AVX2 specialized transpose lowering.

Transform dialect path is not default by IREE codegen. To enable it, we can add `--iree-codegen-llvmcpu-enable-transform-dialect-jit` to compilation flags.

Benchmark results:

src shape | before (us) | after (us) | before/after ratio
-- | -- | -- | --
384x128 | 15.38 | 5.48 | 280.66%
384x512 | 77.23 | 64.6 | 119.55%
384x384 | 53.1 | 39.33 | 135.01%
384x32 | 3.82 | 1.47 | 259.86%
384x2 | 0.402 | 0.323 | 124.46%