@nicolasvasilache 2023-04-15
Nice and easy perf grains with the proper tools, it is great to see this!

General comment: please surface back comments on usability improvements to @ftynse @qcolombet @matthias-springer and @nicolasvasilache so we can help improve the user experience. There is still a long way to go to full automation but it does go through more eyes scrutinizing the abstractions and surfacing pain points.