@MaheshRavishankar 2023-04-15
> Re:
> 
> ```
> Transform dialect path is not default by IREE codegen. To enable it, we can add --iree-codegen-llvmcpu-enable-transform-dialect-jit to compilation flags.
> ```
> 
> Once the refatoring is done, I'd suggest to always use TD and have it tell you whether or not the strategy exists / is good and then fallback to the default codegen otherwise. We can then gradually bring cases up with dedicated flags e.g. `--iree-codegen-llvmcpu-enable-transform-dialect-softmax` that turn individual examples on/off on the TD side without an impact on the rest of the system. As strategies mature, the flags can go and our compiler is improved forever.

Enabling TD by default on CPU side would be a good idea. For things like GEMM that optionality might be useful. Here though, I think things like `pack` are table-stakes. The default should be able to handle this. @hanhanW can this be incorporated into the non-TD path easily?