@hanhanW 2023-04-18
> Enabling TD by default on CPU side would be a good idea. For things like GEMM that optionality might be useful. Here though, I think things like `pack` are table-stakes. The default should be able to handle this. @hanhanW can this be incorporated into the non-TD path easily?

Yes, the default should be able to handle this. I'll add the support to default as well. This is more like a learning of transform dialect to me, and I can use this as a reference to default pipeline.