@hanhanW 2023-05-26
I need some IREE CPU logics and fusion support, so I land it to non-transform dialect path. The below pipeline has similar pipeline to transform dialect spec:

https://github.com/openxla/iree/blob/c63d261951ba5017e9c9564be41cc39e4d278450/compiler/src/iree/compiler/Codegen/LLVMCPU/Passes.cpp#L577-L602

I need it working on fusion cases for other pipelines together, so it's landed in a different form.